+++ Mods ChoGGi/Building Menu Shows Specialists/metadata.lua
@@ -1,0 +1,23 @@
+return PlaceObj("ModDef", {
+	"title", "Building Menu Shows Specialists",
+	"id", "ChoGGi_BuildingMenuShowsSpecialists",
+	"steam_id", "3405723712",
+	"pops_any_uuid", "a9b0dc34-a03f-46be-8bca-94c1ee345c1d",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagInterface", true,
+	"description", [[
+Build menu will show best worker spec.
+
+
+Requested by rigom.works
+]],
+})