+++ Mods ChoGGi/Find Map Locations/metadata.lua
@@ -18,9 +18,9 @@ return PlaceObj("ModDef", {
 	"pops_any_uuid", "b97dbd4e-71cf-4e82-80a0-46dd104133bc",
 	"steam_id", "2453011286",
 	"lua_revision", 1007000, -- Picard
-	"version", 13,
+	"version", 14,
 	"version_major", 1,
-	"version_minor", 3,
+	"version_minor", 4,
 	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {
@@ -33,10 +33,11 @@ Find landing spots with/without a combination of breakthroughs (Extractor AI), r
 Left click an item to move the globe to it.
 
 The first 13 breakthroughs are guaranteed (4 from planetary anomalies, and 9 from ground anomalies). The rest are dependant on story bits/etc.
+Paradox sponsor does add an extra 2-4 (random), so if you play with it then it'll bump the count by 2.
 
 
 Known Issues:
-If you have B&B DLC then the planetary anomalies don't work at all. Either ignore the first four breakthroughs, or use my [url=https://steamcommunity.com/sharedfiles/filedetails/?id=2721921772]Fix Bugs[/url] mod.
+If you have B&B DLC then the planetary anomalies don't work at all. They'll be hidden from search if you don't use my [url=https://steamcommunity.com/sharedfiles/filedetails/?id=2721921772]Fix Bugs[/url] mod.
 If you use the game rules to disable dlc then this won't work, you need to manually remove hpk files from the DLC folder.
 Do not use Tech Variety or Chaos Theory game rules! (they change the breakthrough order).
 This mod requires a mouse and keyboard on xbox.