+++ Mods ChoGGi/Fix Bugs/items.lua
@@ -7,6 +7,36 @@ local mod_options = {
 		"Help", T(302535920011793, "Disable mod without having to see missing mod msg."),
 		"DefaultValue", true,
 	}),
+	-- keep at top
+	PlaceObj("ModItemOptionToggle", {
+		"name", "ColonistsWrongMap",
+		"DisplayName", T(0000, "Colonists Wrong Map Infobar"),
+		"Help", T(0000, [[Having B&B and colonists in the underground will cause some issues (okay having B&B will) with colonists showing up as belonging to the wrong map.
+
+I added it as an option since this is causing issues for a user; it'll lock up on save game loading when it tries to remove some buggy objects.]]),
+		"DefaultValue", false,
+	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "UnevenTerrain",
+		"DisplayName", T(0000, "Uneven Terrain"),
+		"Help", T(0000, [[This calls RefreshBuildableGrid() after a landscaping project completes.
+More info:
+When finishing landscaping it can set some of the surrounding hexes z values (height) to 65535 (also known as UnbuildableZ).
+
+Calling RefreshBuildableGrid() on the map seems to get rid of them without causing any major issues:
+It can mark some hexes as okay to build when they weren't before, but nothing like a cliff side or anything.
+Geoscape Domes will complain about uneven terrain for spires:
+I don't see why the game should be checking for uneven terrain in a dome, so... skip! (at least it should skip them, let me know if you can't build)
+
+If you enable the mod option and notice that you can build on some places you really shouldn't be able to then please let me know :)
+
+If you're bored and want to dig through the funcs in LandscapeFinish() to find out exactly where it's coming from, feel free.
+]]),
+		"DefaultValue", true,
+	}),
+
+
+
 	PlaceObj("ModItemOptionToggle", {
 		"name", "FarmOxygen",
 		"DisplayName", T(0000, "Farm Oxygen"),
@@ -41,32 +71,6 @@ Turn off to enable cheese and disable fix.
 		"DefaultValue", true,
 	}),
 	PlaceObj("ModItemOptionToggle", {
-		"name", "UnevenTerrain",
-		"DisplayName", T(0000, "Uneven Terrain"),
-		"Help", T(0000, [[This calls RefreshBuildableGrid() after a landscaping project completes.
-More info:
-When finishing landscaping it can set some of the surrounding hexes z values (height) to 65535 (also known as UnbuildableZ).
-
-Calling RefreshBuildableGrid() on the map seems to get rid of them without causing any major issues:
-It can mark some hexes as okay to build when they weren't before, but nothing like a cliff side or anything.
-Geoscape Domes will complain about uneven terrain for spires:
-I don't see why the game should be checking for uneven terrain in a dome, so... skip! (at least it should skip them, let me know if you can't build)
-
-If you enable the mod option and notice that you can build on some places you really shouldn't be able to then please let me know :)
-
-If you're bored and want to dig through the funcs in LandscapeFinish() to find out exactly where it's coming from, feel free.
-]]),
-		"DefaultValue", true,
-	}),
---~ 	PlaceObj("ModItemOptionToggle", {
---~ 		"name", "TurnOffUpgrades",
---~ 		"DisplayName", T(0000, "Turn Off Upgrades"),
---~ 		"Help", T(0000, [[Some buildings don't properly turn off their upgrades which causes them to keep their modifiers on.
---~ The "fix" is turning off upgrades when a building is demolished, turned off, malfunctioned (might be annoying, mod option to keep it as is).
---~ ]]),
---~ 		"DefaultValue", false,
---~ 	}),
-	PlaceObj("ModItemOptionToggle", {
 		"name", "SupplyPodSoundEffects",
 		"DisplayName", T(0000, "Supply Pod Sound Effects"),
 		"Help", T(0000, [[Use the Supply Rocket sounds for the Supply Pod (since it doesn't have any).]]),
@@ -78,14 +82,6 @@ If you're bored and want to dig through the funcs in LandscapeFinish() to find o
 		"Help", T(0000, [[If the main menu music keeps playing in-game on new games.
 
 I added it as an option since it removes the fade out.]]),
-		"DefaultValue", true,
-	}),
-	PlaceObj("ModItemOptionToggle", {
-		"name", "ColonistsWrongMap",
-		"DisplayName", T(0000, "Colonists Wrong Map Infobar"),
-		"Help", T(0000, [[Having B&B and colonists in the underground will cause some issues (okay having B&B will) with colonists showing up as belonging to the wrong map.
-
-I added it as an option since this is causing issues for a user; it'll lock up on save game loading when it tries to remove some buggy objects.]]),
 		"DefaultValue", true,
 	}),
 	PlaceObj("ModItemOptionToggle", {